Moved `IncludedBuildInternal` to live with `IncludedBuildRegistry` so that the registry can reference the internal type and used this to simplify the registry API.
